+++ src/models/index.js
@@ -1,0 +1,1 @@
+export { default as CarModel } from './CarModel.js';
